# hero/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /addEvent:
    x-swagger-router-controller: event
    post:
      summary: save event data
      description: save and argument event data
      operationId: addEvent
      parameters:
        - name: eventList
          in: body
          description: list of event JSON objects
          required: true
          schema:
            $ref:  "#/definitions/EventListSchema"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getEventHistory:
    x-swagger-router-controller: event
    get:
      summary: get event history
      description: get a list of all recorded events
      operationId: getEvents
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getMerchantHistory/{merchant}:
    x-swagger-router-controller: merchantHistory
    get:
      operationId: getMerchantHistory
      summary: get merchant history
      description: get merchant history object
      parameters:
        - name: merchant
          in: path
          description: merchant identifier
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MerchantResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  EventListSchema:
    type: array
    items:
      type: object
      required:
        - type
        - merchant
        - user
        - time
        - data
      properties:
        type:
          type: string
        merchant:
          type: string
        user:
          type: string
        time:
          type: string
        data:
          type: object
          properties:
            product:
              type: object
              properties:
                sku_code:
                  type: string
            location:
              type: string
            transaction:
              type: object
              properties:
                order_id:
                  type: string
                subtotal:
                  type: number
                total:
                  type: number
                line_items:
                  type: array
                  items:
                    type: object
                    properties:
                      product:
                        type: object
                        properties:
                          sku_code:
                            type: string
                          price:
                            type: number
                      quantity:
                        type: number
                      subtotal:
                        type: number
  OkResponse:
    required:
      - message
    properties:
      message:
        type: string
  MerchantResponse:
    type: object
    required: 
      - total_events
      - number_of_customers
      - events_summary
    properties:
      total_events:
        type: number
      number_of_customers:
        type: number
      events_summary:
        type: array
        items:
          type: object
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string